refactor(FormGroup): autoBindErrorCallbackRefをFormGroupに直接実装する
autoBindErrorCallbackRefはFormGroup.tsxからしか呼ばれなくなっていた
ため、別ファイルへの切り出しをやめてコンポーネントに直接統合した。
ロジックに変更はない。
